moving glyf_nested_components out of googlefonts directory
