+++ Lib/fontbakery/checks/vendorspecific/googlefonts/glyf.py
@@ -44,39 +44,3 @@ def check_negative_advance_width(ttFont):
                 f" table, which may lead to the advance width to be"
                 f" interpreted as a negative value ({advwidth}).",
             )
-
-
-@check(
-    id="glyf_nested_components",
-    rationale="""
-        There have been bugs rendering variable fonts with nested components.
-        Additionally, some static fonts with nested components have been reported
-        to have rendering and printing issues.
-
-        For more info, see:
-        * https://github.com/fonttools/fontbakery/issues/2961
-        * https://github.com/arrowtype/recursive/issues/412
-    """,
-    conditions=["is_ttf"],
-    proposal="https://github.com/fonttools/fontbakery/issues/2961",
-)
-def check_glyf_nested_components(ttFont, config):
-    """Check glyphs do not have components which are themselves components."""
-    from fontbakery.utils import pretty_print_list
-
-    bad_glyphs = []
-    for glyph_name in ttFont["glyf"].keys():
-        glyph = ttFont["glyf"][glyph_name]
-        if not glyph.isComposite():
-            continue
-        for comp in glyph.components:
-            if ttFont["glyf"][comp.glyphName].isComposite():
-                bad_glyphs.append(glyph_name)
-    if bad_glyphs:
-        formatted_list = "\t* " + pretty_print_list(config, bad_glyphs, sep="\n\t* ")
-        yield FAIL, Message(
-            "found-nested-components",
-            f"The following glyphs have components which"
-            f" themselves are component glyphs:\n"
-            f"{formatted_list}",
-        )
